Create display-substring-matching-elements
Develop a class TelephoneIndex with two String objects as members. One should hold people’s names and the other should hold their phone number. Create an array of objects for TelephoneIndex and do the following:


a) Your program should ask the user to enter a name or the first few characters of a name to search for it in the array.

b) The program should display all of the names that match the user’s input and their corresponding phone numbers.