Create display-address-string
Create a class named Address with the following member variables and methods

street as String
city as String
pincode as integer
country as String
displayAddress()	to display all the details.

Create a main class named AddressMain to include the Main method.

In the main method, obtain the details of the Address by creating an object for the Address class and assign the values to the attributes. Call the method displayAddress() in the Main class to display the values.